requirements.txt
alembic==1.15.2
Flask==3.1.1
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
injector==0.22.0
langchain_core==0.3.60
langchain_openai==0.3.17
pydantic==2.11.4
pytest==8.3.5
python-dotenv==1.1.0
WTForms==3.2.1
